+++ backend/src/test/java/fr/univtln/m1infodid/projet_s2/backend/server/MyResourceTest.java
@@ -1,4 +1,4 @@
-package fr.univtln.projets2.backend.server;
+package fr.univtln.m1infodid.projet_s2.backend.server;
 
 import jakarta.ws.rs.client.Client;
 import jakarta.ws.rs.client.ClientBuilder;
